requirements.txt
Django==1.9
django-cors-headers==1.1.0
django-filter==0.11.0
djangorestframework==3.3.2
MySQL-python==1.2.5
Pillow==3.0.0
wheel==0.24.0
xmltodict==0.9.2
